avancement sur la refactorisation de certains modules infra
