+++ magma-diffusion-impl/src/main/java/fr/insee/rmes/magma/diffusion/queries/parameters/ClassificationRequestParametizer.java
@@ -2,6 +2,6 @@ package fr.insee.rmes.magma.diffusion.queries.parameters;
 
 public record ClassificationRequestParametizer(String nomenclature,
                                                String niveau,
-                                               String code) implements ParametersForQuery<ClassificationRequestParametizer> {
+                                               String code) implements ParametersForQueryDiffusion<ClassificationRequestParametizer> {
 
 }
