+++ magma-diffusion-impl/src/main/java/fr/insee/rmes/magma/diffusion/queries/parameters/ConceptsRequestParametizer.java
@@ -1,7 +1,7 @@
 package fr.insee.rmes.magma.diffusion.queries.parameters;
 
 
-public record ConceptsRequestParametizer(String id, String label, String uriConcept) implements ParametersForQuery<ConceptsRequestParametizer>{
+public record ConceptsRequestParametizer(String id, String label, String uriConcept) implements ParametersForQueryDiffusion<ConceptsRequestParametizer> {
 
     public ConceptsRequestParametizer(String id,
                                   String label){
