+++ magma-diffusion-impl/src/main/java/fr/insee/rmes/magma/diffusion/queries/parameters/ParameterValueDecoderDiffusion.java
@@ -1,13 +1,12 @@
 package fr.insee.rmes.magma.diffusion.queries.parameters;
 
 import fr.insee.rmes.magma.diffusion.model.*;
+import fr.insee.rmes.magma.queries.parameters.ParameterValueDecoder;
 
-import java.lang.reflect.RecordComponent;
 import java.time.LocalDate;
-import java.util.function.Function;
 
 @FunctionalInterface
-interface ParameterValueDecoder<T> {
+interface ParameterValueDecoderDiffusion<T> extends ParameterValueDecoder<T> {
 
     String STRING_CLASS = "java.lang.String";
     String BOOLEAN_CLASS = "boolean";
@@ -38,7 +37,7 @@ interface ParameterValueDecoder<T> {
     String ENUM_DESCENDANTS_ZONEDEMPLOI_CLASS = "fr.insee.rmes.magma.diffusion.model.TypeEnumDescendantsZoneDEmploi";
     String ENUM_TERRITOIRE_LIE = "fr.insee.rmes.magma.diffusion.model.TypeEnum";
 
-    static <U> ParameterValueDecoder<U> of(Class<U> type) {
+    static <U> ParameterValueDecoderDiffusion<U> of(Class<U> type) {
         return switch (type.getName()) {
             case BOOLEAN_CLASS -> bool -> (boolean) bool ? "true" : "false";
             case CLASS_CLASS -> clazz -> ((Class<?>) clazz).getSimpleName();
@@ -95,36 +94,6 @@ interface ParameterValueDecoder<T> {
             case String ignored when Enum.class.isAssignableFrom(type) -> simpleEnum -> ((Enum<?>) simpleEnum).name();
             default -> throw new IllegalArgumentException("Unsupported type: " + type.getName());
         };
-
-    }
-
-    String decode(T value);
-
-    /**
-     * Class for custom decoder in {@link ParametersForQuery} records such as {@link AscendantsDescendantsRequestParametizer} which
-     * need to customize decoder for a type which has yet a standard decoder in the children of ParameterValueDecoder
-     * <p>
-     * For example {@link AscendantsDescendantsRequestParametizer} needs a custom treatment for Strings for attribute <code>filtreNom</code>.
-     * So it overrides the method {@link  ParametersForQuery#findParameterValueDecoder(RecordComponent)} and for the attribute
-     * <code>filtreNom</code>, it returns its custom decoder with this kind of code :
-     * <code>
-     * if ("filtreNom".equals(recordComponent.getName())){
-     * return new ParameterValueDecoder.DelegaterDecoder<String>(value -> value ==null?"*": value);
-     * }
-     * </code>
-     * <p>
-     * The DelegaterDecoder is never be returned by
-     * { @link fr.insee.rmes.magma.diffusion.queries.parameters.ParameterValueDecoder#of(java.lang.Class)} and can only be
-     * used when explicitly instanced in a method such as {@link AscendantsDescendantsRequestParametizer#findParameterValueDecoder(RecordComponent)}
-     *
-     * @param delegatedDecoder : a function applied to decode the value
-     * @param <U>              : the type for which the instance will decode
-     */
-    record DelegaterDecoder<U>(Function<U, String> delegatedDecoder) implements ParameterValueDecoder<U> {
-        @Override
-        public String decode(U value) {
-            return delegatedDecoder.apply(value);
-        }
     }
 
 }
