+++ magma-diffusion-impl/src/main/java/fr/insee/rmes/magma/diffusion/queries/parameters/TerritoireEtoileRequestParametizer.java
@@ -8,7 +8,7 @@ public record TerritoireEtoileRequestParametizer(String code,
                                                  Class<?> typeOrigine,
                                                  String filtreNom,
                                                  String chefLieu,
-                                                 boolean com) implements ParametersForQuery<TerritoireRequestParametizer> {
+                                                 boolean com) implements ParametersForQueryDiffusion<TerritoireRequestParametizer> {
 
     //for geo/departements, arrondissements, arrondissementsMunicipaux,
     public TerritoireEtoileRequestParametizer(String date,
@@ -36,8 +36,8 @@ public record TerritoireEtoileRequestParametizer(String code,
 
 
     @Override
-    public ParameterValueDecoder<?> findParameterValueDecoder(RecordComponent recordComponent) {
-        return ParametersForQuery.super.findParameterValueDecoder(recordComponent);
+    public ParameterValueDecoderDiffusion<?> findParameterValueDecoder(RecordComponent recordComponent) {
+        return ParametersForQueryDiffusion.super.findParameterValueDecoder(recordComponent);
     }
 
 }
