+++ magma-diffusion-impl/src/main/java/fr/insee/rmes/magma/diffusion/queries/parameters/TerritoireRequestParametizer.java
@@ -10,7 +10,7 @@ public record TerritoireRequestParametizer(String code,
                                            Class<?> typeOrigine,
                                            String filtreNom,
                                            String chefLieu,
-                                           boolean com) implements ParametersForQuery<TerritoireRequestParametizer> {
+                                           boolean com) implements ParametersForQueryDiffusion<TerritoireRequestParametizer> {
 
     //for geo/departement/{code} and geo/region/{code}
     public TerritoireRequestParametizer(String code,
@@ -54,8 +54,8 @@ public record TerritoireRequestParametizer(String code,
     }
 
     @Override
-    public ParameterValueDecoder<?> findParameterValueDecoder(RecordComponent recordComponent) {
-        return ParametersForQuery.super.findParameterValueDecoder(recordComponent);
+    public ParameterValueDecoderDiffusion<?> findParameterValueDecoder(RecordComponent recordComponent) {
+        return ParametersForQueryDiffusion.super.findParameterValueDecoder(recordComponent);
     }
